+++ application/packages/cron/src/fetch-articles.ts
@@ -1,36 +1,10 @@
 import { wrapAsyncCall } from '@trend-diary/common/result'
-import { articles } from '@trend-diary/datastore/drizzle-orm/schema'
-import getRdbClient, { wrapDbCall } from '@trend-diary/datastore/rdb'
 import type { ArticleMedia } from '@trend-diary/domain/article/media'
-import { inArray } from 'drizzle-orm'
-import { err, ok, type Result } from 'neverthrow'
+import { err, type Result } from 'neverthrow'
 import Parser from 'rss-parser'
+import { type CronEnv, storeArticles } from './store-articles'
 
-type CronEnv = {
-  DB: D1Database
-  LOG_LEVEL?: import('@trend-diary/common/logger').LogLevel
-}
-
-type D1Database = import('@cloudflare/workers-types').D1Database
-
-type FeedItem = {
-  title: string
-  author: string
-  description: string
-  url: string
-}
-
-const MAX_LENGTH = {
-  media: 10,
-  title: 100,
-  author: 30,
-  description: 1024,
-}
 const HATENA_FALLBACK_AUTHOR = 'はてなブックマーク'
-
-function truncateByCodePoint(text: string, maxLength: number): string {
-  return [...text].slice(0, maxLength).join('')
-}
 
 function pickNonEmpty(...candidates: Array<string | undefined>): string | undefined {
   for (const candidate of candidates) {
@@ -56,81 +30,6 @@ async function fetchRssFeed<T>(url: string): Promise<Result<T[], Error>> {
     const feed = await parser.parseString(xml)
     return feed.items
   })
-}
-
-async function storeArticles(
-  media: ArticleMedia,
-  items: FeedItem[],
-  env: CronEnv,
-): Promise<Result<number, Error>> {
-  const db = getRdbClient(env.DB)
-  if (items.length === 0) return ok(0)
-
-  const normalized = items.map((item) => ({
-    media: truncateByCodePoint(media, MAX_LENGTH.media),
-    title: truncateByCodePoint(item.title, MAX_LENGTH.title),
-    author: truncateByCodePoint(item.author, MAX_LENGTH.author),
-    description: truncateByCodePoint(item.description, MAX_LENGTH.description),
-    url: item.url,
-  }))
-
-  const existingResult = await wrapDbCall(() =>
-    db
-      .select({ url: articles.url })
-      .from(articles)
-      .where(
-        inArray(
-          articles.url,
-          normalized.map((item) => item.url),
-        ),
-      ),
-  )
-  if (existingResult.isErr()) {
-    return err(existingResult.error)
-  }
-
-  const existingUrlSet = new Set(existingResult.value.map((item) => item.url))
-  const feedUrlSet = new Set<string>()
-  const uniqueNormalized: typeof normalized = []
-  for (const article of normalized) {
-    if (feedUrlSet.has(article.url)) continue
-    feedUrlSet.add(article.url)
-    uniqueNormalized.push(article)
-  }
-  const toInsert = uniqueNormalized.filter((item) => !existingUrlSet.has(item.url))
-
-  if (toInsert.length === 0) return ok(0)
-
-  let insertedCount = 0
-  for (const article of toInsert) {
-    // INFO: D1互換のため記事は1件ずつ保存する
-    const insertResult = await wrapDbCall(() => db.insert(articles).values(article))
-    if (insertResult.isErr()) {
-      if (isUniqueConstraintError(insertResult.error)) continue
-      return err(insertResult.error)
-    }
-    insertedCount += 1
-  }
-
-  return ok(insertedCount)
-}
-
-// Drizzle はドライバエラーを DrizzleQueryError でラップし元エラーを cause に格納するため、
-// cause チェーンを辿って 'UNIQUE constraint failed'(D1/SQLite共通)を探す。
-function isUniqueConstraintError(error: unknown): boolean {
-  let current: unknown = error
-  for (let depth = 0; depth < 5 && current != null; depth += 1) {
-    if (typeof current !== 'object') break
-    const candidate = current as { message?: unknown; cause?: unknown }
-    if (
-      typeof candidate.message === 'string' &&
-      candidate.message.includes('UNIQUE constraint failed')
-    ) {
-      return true
-    }
-    current = candidate.cause
-  }
-  return false
 }
 
 export async function fetchQiitaArticles(env: CronEnv): Promise<Result<number, Error>> {
